inital work on multi-toolenv for parallel tasks
